### 2025-10-13 1e66387
Update FAQ, installation documentation, and notebooks
