### 2025-10-13 95c1355
Update FAQ, installation documentation, and notebooks
