### 2025-10-14 fe57fc5
Refactor type hints using annotations module. This is for migration to Python 3.14 in the future with deferred annotation.
