### 2025-10-14 7c5f942
Add string representation call for ImpactModel in test_predict_after_cleanup
