+++ tests/test_predict.py
@@ -148,6 +148,7 @@ def test_predict_after_cleanup(
     with pytest.warns(UserWarning, match=msg):
         im.predict(X=X, batch_size=len(X) // 2, progress=False)
     temp_dir_before = im.temp_dir
+    str(im)
     im.cleanup()
     with pytest.warns(UserWarning, match=msg):
         im.predict(X=X, batch_size=len(X) // 2, progress=False)
