### 2025-10-14 72776c0
Update changelog for v0.8.0: add string representation methods to ImpactModel and include sample_sites in KernelSpec
Fixes #118
Fixes #125
