### 2025-10-14 d623649
Fix string representation formatting in ImpactModel for Python 3.11
