### 2025-10-23 7976172
Update pre-commit hooks to latest versions and adjust Sphinx configuration for MLflow integration
