### 2025-10-23 45a146d
Update pre-commit hooks to latest versions and adjust Sphinx configuration for MLflow integration
