### 2025-10-24 3dd636c
Replaced deprecated `jax.experimental.shard_map.shard_map` with `jax.shard_map`.
Fix #128.
