### 2025-10-24 35fa5d9
Replaced deprecated `jax.experimental.shard_map.shard_map` with `jax.shard_map`.
Fix #128.
