### 2025-10-24 1da240c
Improve logging during writer thread shutdown
Fixes #130
