+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "aimz"
-version = "0.8.0"
+version = "0.8.1"
 description = "Scalable probabilistic impact modeling"
 readme = "README.md"
 license = "Apache-2.0"
@@ -15,11 +15,11 @@ authors = [
 maintainers = [{ name = "Eunseop Kim", email = "markean@pm.me" }]
 requires-python = ">=3.11"
 dependencies = [
-    "dask>=2025.5",
-    "jax>=0.7",
+    "dask>=2025.7",
+    "jax>=0.8",
     "numpyro>=0.19.0",
     "scikit-learn>=1.6.1",
-    "xarray>=2025.4",
+    "xarray>=2025.7",
     "zarr>=3,<4",
 ]
 classifiers = [
@@ -36,15 +36,15 @@ dev = ["pre-commit>=4.3", "pytest>=8.4", "pytest-cov>=7", "ruff>=0.14"]
 gpu = ["jax[cuda13]>=0.7"]
 mlflow = ["mlflow>=3"]
 docs = [
-    "sphinx>=8.2",
-    "myst-parser>=4.0",
+    "arviz-plots[matplotlib]>=0.6",
+    "flax>=0.12",
     "jupyter-sphinx>=0.5",
+    "mlflow>=3",
+    "myst-parser>=4.0",
+    "pydata-sphinx-theme>=0.16",
+    "sphinx>=8.2",
     "sphinx-copybutton>=0.5",
     "sphinx-design>=0.6",
-    "pydata-sphinx-theme>=0.16",
-    "flax>=0.12",
-    "mlflow>=3",
-    "arviz-plots[matplotlib]>=0.6",
 ]
 
 [project.urls]
