### 2025-11-10 4e94652
Update dependencies and add documentation source files. Bump to v0.8.2.dev0.
