+++ tests/test_sample.py
@@ -17,12 +17,12 @@
 from typing import TYPE_CHECKING
 
 import pytest
-from conftest import lm
 from jax import random
 from jax.typing import ArrayLike
 from numpyro.infer import MCMC
 
 from aimz import ImpactModel
+from tests.conftest import lm
 
 if TYPE_CHECKING:
     from numpyro.infer import SVI
