+++ tests/test_predict.py
@@ -18,7 +18,6 @@ from tempfile import TemporaryDirectory
 
 import numpyro.distributions as dist
 import pytest
-from conftest import latent_variable_model, lm
 from jax import random
 from jax.typing import ArrayLike
 from numpyro import sample
@@ -28,6 +27,7 @@ from numpyro.optim import Adam
 
 from aimz import ImpactModel
 from aimz._exceptions import NotFittedError
+from tests.conftest import latent_variable_model, lm
 
 
 def test_model_not_fitted() -> None:
