+++ tests/test_log_likelihood.py
@@ -15,7 +15,6 @@
 """Tests for the `.log_likelihood()` method."""
 
 import pytest
-from conftest import lm
 from jax import random
 from jax.typing import ArrayLike
 from numpyro.infer import SVI, Trace_ELBO
@@ -24,6 +23,7 @@ from numpyro.optim import Adam
 
 from aimz import ImpactModel
 from aimz._exceptions import NotFittedError
+from tests.conftest import lm
 
 
 def test_model_not_fitted() -> None:
