+++ tests/test_prng_state.py
@@ -16,12 +16,12 @@
 
 import jax.numpy as jnp
 import pytest
-from conftest import lm
 from jax import random
 from jax.typing import ArrayLike
 from numpyro.infer import SVI
 
 from aimz import ImpactModel
+from tests.conftest import lm
 
 
 @pytest.mark.parametrize("vi", [lm], indirect=True)
