+++ tests/test_init.py
@@ -15,10 +15,10 @@
 """Tests for initializing the `ImpactModel` class."""
 
 import pytest
-from conftest import lm
 from jax import random
 
 from aimz import ImpactModel
+from tests.conftest import lm
 
 
 def test_unsupported_inference_method() -> None:
