+++ tests/__init__.py
@@ -1,1 +1,5 @@
 """Test suite for aimz."""
+
+from tests.conftest import latent_variable_model, lm, lm_with_kwargs_array
+
+__all__ = ["latent_variable_model", "lm", "lm_with_kwargs_array"]
