### 2025-11-11 9bb1031
Add class-level cleanup method
Fixes #136
