### 2025-11-11 6f85ba9
Add class-level cleanup method
Fixes #136
