### 2025-11-14 b6ed2d3
Update pre-commit hooks and changelog; change output subdirectory naming convention to include caller name
Fix #138
