### 2025-11-14 3eefa90
Update pre-commit hooks and changelog; change output subdirectory naming convention to include caller name
Fix #138
