### 2025-11-14 b0f0c9e
Lower logging level for exceptions during temporary directory cleanup from 'exception' to 'debug'.
Fix #140.
