### 2025-11-17 ca4b0b2
Bump version to 0.9.0 and update changelog; add installation instructions to README
