### 2025-11-17 252432c
Bump version to 0.9.0 and update changelog; add installation instructions to README
