+++ .github/workflows/ruff.yaml
@@ -15,7 +15,7 @@ jobs:
 
     steps:
       - name: Checkout repository
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6
 
       - name: ruff-action
         uses: astral-sh/ruff-action@v3
