### 2025-12-08 a3aa269
fix forward sampler explicit sharding compatibility
