### 2025-12-08 6d1ad58
fix forward sampler explicit sharding compatibility
