### 2025-12-08 d2983df
Set check_vma parameter to False in _create_sharded_sampler function
