### 2025-12-08 61f051c
bump version to 0.9.1 in pyproject.toml and uv.lock
