+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file and are best
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.9.1](https://github.com/markean/aimz/releases/tag/v0.9.1) - 2025-12-08
+
+### Fixed
+
+- Fixed {func}`jax.shard_map` closure error for sharded `rng_key` in parallelism methods when using JAX 0.8 and newer versions ([#140](https://github.com/markean/aimz/issues/140)).
+
 ## [v0.9.0](https://github.com/markean/aimz/releases/tag/v0.9.0) - 2025-11-16
 
 ### Added
