### 2025-12-08 c7eeee6
bump version to 0.9.1 in pyproject.toml and uv.lock
