### 2025-12-31 #9183 @rich2355
Yes, I am happy to take this!
