### 2026-01-08 #9738 @editorialbot
[REVIEW]: aimz: Scalable probabilistic impact modeling
**Submitting author:** @markean (<a href="http://orcid.org/0009-0000-2138-788X">Eunseop Kim</a>)
**Repository:** https://github.com/markean/aimz
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.7.0
