### 2026-01-20 #9738 @ankurankan
> **Paper file info**:
>
> 📄 Wordcount for `paper.md` is **508**
>
> ✅ The paper includes a `Statement of need` section
>
> ```
> 🔴 Failed to discover a `Software Design` section in paper
>
> 🔴 Failed to discover a `Research Impact Statement` section in paper
>
> 🔴 Failed to discover an `AI usage disclosure` section in paper
> ```

 @markean I think the paper is still missing some of the required sections. I will wait for it before I go through the checklist.
