### 2026-01-20 #9738 @markean
> > **Paper file info**:
> > 📄 Wordcount for `paper.md` is **508**
> > ✅ The paper includes a `Statement of need` section
> > ```
> > 🔴 Failed to discover a `Software Design` section in paper
> >
> > 🔴 Failed to discover a `Research Impact Statement` section in paper
> >
> > 🔴 Failed to discover an `AI usage disclosure` section in paper
> > ```
>
> [@markean](https://github.com/markean) I think the paper is still missing some of the required sections. I will wait for it before I go through the checklist.

@rich2355 - I noticed that the JOSS paper format requirements have been updated, and some other submissions have the pre-2026-submissions tag. Could you provide guidance on the review process and any steps I should take on my end?
