### 2026-01-22 5f99f7a
Update type annotations and dependencies, addressing issues reported by the type checker.
