+++ aimz/model/impact_model.py
@@ -89,7 +89,7 @@ class ImpactModel(BaseModel):
     def __init__(
         self,
         kernel: Callable,
-        rng_key: ArrayLike,
+        rng_key: Array,
         inference: SVI | MCMC,
         *,
         param_input: str = "X",
@@ -99,7 +99,7 @@ class ImpactModel(BaseModel):
 
         Args:
             kernel: A probabilistic model with NumPyro primitives.
-            rng_key (ArrayLike): A pseudo-random number generator key.
+            rng_key: A pseudo-random number generator key.
             inference: An inference method supported by NumPyro, such as an instance of
                 :external:class:`~numpyro.infer.svi.SVI` or
                 :external:class:`~numpyro.infer.mcmc.MCMC`.
@@ -240,7 +240,7 @@ class ImpactModel(BaseModel):
         return self._kernel_spec
 
     @property
-    def rng_key(self) -> ArrayLike:
+    def rng_key(self) -> Array:
         """Pseudo-random number generator key."""
         return self._rng_key
 
@@ -403,7 +403,7 @@ class ImpactModel(BaseModel):
     def _sample_and_write(
         self,
         num_samples: int,
-        rng_key: ArrayLike,
+        rng_key: Array,
         return_sites: tuple[str, ...],
         output_dir: Path,
         kernel: Callable,
@@ -519,7 +519,7 @@ class ImpactModel(BaseModel):
         X: ArrayLike,
         *,
         num_samples: int = 1000,
-        rng_key: ArrayLike | None = None,
+        rng_key: Array | None = None,
         return_sites: str | Iterable[str] | None = None,
         return_datatree: bool = True,
         **kwargs: object,
@@ -529,8 +529,8 @@ class ImpactModel(BaseModel):
         Args:
             X (ArrayLike): Input data with shape ``(n_samples_X, n_features)``.
             num_samples: The number of samples to draw.
-            rng_key (ArrayLike | None): A pseudo-random number generator key. By
-                default, an internal key is used and split as needed.
+            rng_key: A pseudo-random number generator key. By default, an internal key
+                is used and split as needed.
             return_sites: Names of variables (sites) to return. If ``None``, samples
                 :attr:`~aimz.ImpactModel.param_output` and deterministic sites.
             return_datatree: If ``True``, return a :external:class:`~xarray.DataTree`;
@@ -585,7 +585,7 @@ class ImpactModel(BaseModel):
         X: ArrayLike,
         *,
         num_samples: int = 1000,
-        rng_key: ArrayLike | None = None,
+        rng_key: Array | None = None,
         return_sites: str | Iterable[str] | None = None,
         batch_size: int | None = None,
         output_dir: str | Path | None = None,
@@ -600,8 +600,8 @@ class ImpactModel(BaseModel):
         Args:
             X (ArrayLike): Input data with shape ``(n_samples_X, n_features)``.
             num_samples: The number of samples to draw.
-            rng_key (ArrayLike | None): A pseudo-random number generator key. By
-                default, an internal key is used and split as needed.
+            rng_key: A pseudo-random number generator key. By default, an internal key
+                is used and split as needed.
             return_sites: Names of variables (sites) to return. If ``None``, samples
                 :attr:`~aimz.ImpactModel.param_output` and deterministic sites.
             batch_size: The batch size for data loading during prior predictive
@@ -725,7 +725,7 @@ class ImpactModel(BaseModel):
         self,
         *,
         num_samples: int = 1000,
-        rng_key: ArrayLike | None = None,
+        rng_key: Array | None = None,
         return_sites: str | Iterable[str] | None = None,
         return_datatree: bool = True,
         **kwargs: object,
[…]
