### 2026-01-22 2866f35
Update type annotations and dependencies, addressing issues reported by the type checker.
