+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "aimz"
-version = "0.9.1"
+version = "0.9.2.dev0"
 description = "Scalable probabilistic impact modeling"
 readme = "README.md"
 license = "Apache-2.0"
@@ -26,6 +26,8 @@ classifiers = [
     "Development Status :: 3 - Alpha",
     "Intended Audience :: Science/Research",
     "Topic :: Scientific/Engineering",
+    'Programming Language :: Python :: 3',
+    'Programming Language :: Python :: 3 :: Only',
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
@@ -33,18 +35,18 @@ classifiers = [
 
 [project.optional-dependencies]
 dev = ["pre-commit>=4.3", "pytest>=9", "pytest-cov>=7", "ruff>=0.14"]
-gpu = ["jax[cuda13]>=0.7"]
+gpu = ["jax[cuda13]>=0.8"]
 mlflow = ["mlflow>=3"]
 docs = [
     "arviz-plots[matplotlib]>=0.6",
     "flax>=0.12",
     "jupyter-sphinx>=0.5",
     "mlflow>=3",
-    "myst-parser>=4.0",
+    "myst-parser>=5",
     "pydata-sphinx-theme>=0.16",
     "sphinx>=8.2",
     "sphinx-copybutton>=0.5",
-    "sphinx-design>=0.6",
+    "sphinx-design>=0.7",
 ]
 
 [project.urls]
@@ -71,9 +73,8 @@ per-file-ignores = { "aimz/model/_core.py" = [
     "PLR0913", # too-many-arguments
     "PLC0415", # import-outside-top-level
 ], "docs*" = [
-    # "INP001", # implicit-namespace-package
-    "A001",   # builtin-variable-shadowing
-    "F401",   # unused-import
+    "A001", # builtin-variable-shadowing
+    "F401", # unused-import
 ], "tests*" = [
     "ARG001", # unused-function-argument
     "S101",   # assert
