### 2026-01-22 6451648
Update mlflow version to 3.5 in pyproject.toml and uv.lock
