+++ paper/paper.md
@@ -40,4 +40,39 @@ Existing impact or uplift modeling libraries (e.g., domain‑specific frameworks
 `aimz` instead pursues generality—accepting arbitrary NumPyro model functions and multiple inference strategies—while still avoiding boilerplate through streamed predictive simulation, structured outputs, and intervention orchestration.
 By elevating scalable posterior predictive simulation and intervention effect estimation to first-class capabilities, `aimz` lowers the barrier between exploratory probabilistic modeling and production-grade Bayesian impact analysis.
 
-## References
+# Software design
+
+`aimz` is designed to support scalable Bayesian analyses in applied settings, allowing users to iterate quickly across model specifications, inference settings, and intervention scenarios, as well as to handle large datasets and produce reproducible artifacts.
+The library is built on NumPyro to leverage its modeling flexibility and JAX’s accelerator-native ecosystem and composable program transformations (e.g., JIT, vectorization, and sharded execution), which are well suited to impact modeling workflows where posterior predictive simulation and scenario evaluation often dominate overall runtime relative to model fitting.
+
+`aimz` prioritizes performance as a key consideration.
+Predictive sampling and effect estimation are organized around JIT-accelerated execution, data-parallel sharding, and streaming results to disk-backed, chunked storage, enabling sustained throughput for large simulation workloads without requiring all draws to be retained in memory.
+This architectural choice comes with trade-offs: only models that are compatible with sharded execution can fully leverage these optimizations.
+Additional orchestration around execution and I/O is required, which `aimz` manages through a small set of high-level methods, enabling users to focus on analysis while maintaining a consistent interface.
+At the user interface level, `aimz` adopts an estimator-like API centered on the `ImpactModel` class, which takes a NumPyro model function as its “kernel” (the primary user-specified argument) and exposes familiar estimator methods while still supporting a wide range of model specifications.
+
+`aimz` is also designed for integration: results are materialized as structured artifacts (e.g., Xarray objects backed by Zarr stores) and can optionally be logged via MLflow for experiment tracking and lineage.
+The combination of a stable, method-based interface and standardized outputs makes `aimz` well suited for AI-enabled workflows, where agentic tools can invoke a small set of operations and reliably consume the resulting artifacts.
+
+# Research impact statement
+
+`aimz` enables a class of Bayesian modeling and probabilistic analyses that are otherwise costly to implement and difficult to reproduce at scale: including fitting flexible probabilistic models, generating large posterior predictive simulations, and estimating intervention effects under explicit structural modifications.
+Its research impact is therefore primarily infrastructural---lowering the engineering barrier to rigorous uncertainty-aware effect estimation---while remaining general enough to support diverse model specifications and inference strategies.
+
+Within Eli Lilly and Company, `aimz` has supported internal analytics workflows that require scalable posterior and posterior predictive sampling, consistent output structures, and experiment traceability.
+In this setting, the library has reduced duplicated “glue” code for streaming predictive draws, coordinating intervention scenarios, and tracking model lineage, thereby improving iteration speed and reproducibility across analyses.
+Its stable estimator-like interface and structured, reproducible outputs also make it well suited for integration with AI-enabled workflows, where agentic tools can leverage standardized artifacts for downstream tasks.
+
+Beyond immediate use, `aimz` is designed as reusable research infrastructure: it exposes arbitrary NumPyro model functions through a stable estimator-like interface and standardizes artifacts (samples, predictions, and effect estimates) in formats that integrate cleanly with the broader scientific Python ecosystem (e.g., Xarray/Zarr) and with experiment tracking via MLflow.
[…]
