+++ paper/paper.bib
@@ -124,6 +124,7 @@
   author  = {Gufeng Zhou and Bernardo Lares and Igor Skokan and Leonel Sentana},
   url     = {https://github.com/facebookexperimental/Robyn},
   version = {3.12.0},
+  doi     = {10.32614/cran.package.robyn},
   year    = {2024}
 }
 
