+++ paper/paper.bib
@@ -90,16 +90,6 @@
   url     = {https://arviz-devs.github.io/arviz/}
 }
 
-@article{scikit-learn,
-  author  = {Pedregosa, Fabian and Varoquaux, Ga{"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and Vanderplas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, Matthieu and Duchesnay, {\'E}douard},
-  title   = {Scikit-learn: Machine learning in {P}ython},
-  journal = {Journal of Machine Learning Research},
-  year    = {2011},
-  volume  = {12},
-  pages   = {2825--2830},
-  url     = {https://jmlr.org/papers/v12/pedregosa11a.html}
-}
-
 @article{pymc2023,
   title     = {PyMC: A Modern and Comprehensive Probabilistic Programming Framework in Python},
   author    = {Abril-Pla Oriol and Andreani Virgile and Carroll Colin and Dong Larry and Fonnesbeck Christopher J. and Kochurov Maxim and Kumar Ravin and Lao Jupeng and Luhmann Christian C. and Martin Osvaldo A. and Osthege Michael and Vieira Ricardo and Wiecki Thomas and Zinkov Robert},
@@ -117,6 +107,14 @@
   url     = {https://github.com/google/meridian},
   version = {1.2.1},
   year    = {2025}
+}
+
+@software{causalpy_github,
+  title   = {CausalPy: Causal inference for quasi-experiments in Python},
+  author  = {{PyMC Labs}},
+  url     = {https://github.com/pymc-labs/CausalPy},
+  version = {0.7.0},
+  year    = {2026}
 }
 
 @software{robyn,
