### 2026-01-22 de62fc0
Add state of the field section to the paper.md and refactor other sections accordingly. Update paper.bib with CausalPy citation and remove scikit-learn citation.
