### 2026-01-22 #9738 @markean
Thank you for the guidance, @crvernon and @rich2355. I’ve updated the paper to reflect the new JOSS requirements.

@DanWaxman and @ankurankan, the paper is now ready for your review.
