### 2026-02-20 markean/aimz#161 @DanWaxman
[JOSS Review] Paper Comments
Hi, I'm a reviewer for the corresponding [JOSS submission](https://github.com/openjournals/joss-reviews/issues/9738). Below are comments regarding the paper:

1. On lines 49-50, it is written "focusing on the high-throughput engineering needed to efficiently handle large-scale probabilistic simulations **and persist results** as structured, reusable artifacts." This reads awkwardly to me -- I think it reads much better if "persist" is replaced with "present".
2. The criterion in the reviewing checklist for the summary is that the section "has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?" While I think the summary is relatively clear for the target audience, I'm not sure it accomplishes the criterion for a "diverse, non-specialist audience." Namely, knowledge of probabilistic impact modeling is assumed, as are other technical details (e.g., chunked Zarr stores). I would recommend more gently explaining what probabilistic impact modeling is, then move details about things like SVI or chunked Zarr stores to the body -- for the summary, it should be efficient to say that `aimz` exposes advanced inference methods and efficient serialization, etc.
3. For references, I think it's preferable to use the actual DOI instead of a SemanticScholar link for Zaharia et al. (2018).
