### 2026-02-21 markean/aimz#162 @DanWaxman
[JOSS Review] Clarify What Models Are Compatible
Hi, I'm a reviewer for the corresponding [JOSS Review](https://github.com/openjournals/joss-reviews/issues/9738). I had a question about model compatibility.

In the FAQs, it's stated:
> Can I use aimz with any [NumPyro](https://num.pyro.ai/) model?
> No. Most conventional SVI / MCMC models with global latents and plate-based structure work, but there are limitations:
>
> Use of the control‑flow primitive: [scan()](https://num.pyro.ai/en/stable/primitives.html#numpyro.contrib.control_flow.scan) (or deeply nested dynamic control flow) – only partial support.
>
> Complex per‑example local latent structure (e.g., very large / nested plates) – only partial support.
>
> Models that rely on shapes incompatible with sharded / batched execution may require refactoring.
>
> If a model traces successfully once but fails in batched prediction, try the *_on_batch variants or simplify local latent structure. We plan to broaden coverage—if you hit an unsupported pattern (ideally with a minimal reproducible example), please open an issue or submit a PR.

I think it's quite reasonable not to have full support for arbitrary NumPyro models, but I think it would be useful to be more explicit here. E.g., why are deeply nested plates only partially supported? In what way do they lack support? At what level of nesting is this a problem (4? 20?)? These could be very useful if you're writing a model that you want to be aimz-compatible.
