### 2026-02-21 markean/aimz#163 @DanWaxman
[JOSS Review] Make Impact Modeling Examples More Prominent
Hi, once again part of my [JOSS Review](https://github.com/openjournals/joss-reviews/issues/9738). I'm having some trouble checking off the following two options from some documentation reasons:
> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

While the paper and FAQ stress that a primary interest is in probabilistic impact modeling, there's only [one example of interventions](https://aimz.readthedocs.io/latest/user_guide/intervention.html), somewhat deep into example usage. Given this is a primary focus of the library, there would preferably be several different examples -- including interpretation -- more prominently in the documentation.
