### 2026-02-21 markean/aimz#163 @DanWaxman
To this end, I have some trouble with the "statement of need," in that what is meant by "probabilistic impact modeling" is never very clearly defined. I think a quality tutorial could go a long ways in fixing all of these.
