### 2026-02-21 #9738 @DanWaxman
Hi, I've completed my review and raised issues for parts I couldn't check in my checklist. Namely:
1. In https://github.com/markean/aimz/issues/161, I left some paper comments (for items: Paper -> Summary and Paper -> References).
2. In https://github.com/markean/aimz/issues/162, I left comments about clarifying valid kernels (for item: Documentation -> Functionality Documentation).
3. In https://github.com/markean/aimz/issues/163, I request some more explicit documentation and examples (for items: Documentation -> A Statement of Need, Documentation -> Example Usage, and Documentation -> Functionality Documentation).

---

Some remaining things I have left unchecked that I mainly want to ask the editor @rich2355 for some guidance on:
1. For item: Development -> Collaborative Effort, this is plainly not satisfied, as the library is almost entirely written by a single author.
2. For item: Paper -> Research Impact Statement, I'm unsure the extent to which the paper satisfies this. The current guidance on the JOSS website is:
> Research impact statement (required)[](https://joss.readthedocs.io/en/latest/review_criteria.html#research-impact-statement-required)
> The paper must provide evidence of either:
>
> Realized impact:
>
> Publications using the software
>
> Evidence of external use and adoption
>
> Integration with other research tools or workflows
>
> OR credible near-term significance:
>
> Benchmark results demonstrating improvements
>
> Reproducible research materials showing capabilities
>
> Community-readiness signals (e.g., requests from other groups, presentations at relevant venues)
>
> The evidence should be compelling and specific, not aspirational.
>
> Good: Multiple citations, documented external users, or strong benchmark results with reproducible materials
> OK: Some evidence of use beyond the authors’ group or solid benchmarks
> Bad (not acceptable): Only statements about potential future use without concrete evidence

There are a few different pieces of evidence provided. The first is a somewhat "internal user" perspective that I can't personally verify:
> Within Eli Lilly and Company, aimz has supported internal analytics workflows that require scalable posterior and posterior predictive sampling, consistent output structures, and experiment traceability. In this setting, the library has reduced duplicated “glue” code for streaming predictive draws, coordinating intervention scenarios, and tracking model lineage, thereby improving iteration speed and reproducibility across analyses. Its stable estimator-like interface and structured, reproducible outputs also make it well suited for integration with AI-enabled workflows, where agentic tools can leverage standardized artifacts for downstream tasks.

The second is some potential future use:
> Beyond immediate use, aimz is designed as reusable research infrastructure: it exposes arbitrary NumPyro model functions through a stable estimator-like interface and standardizes artifacts (samples, predictions, and effect estimates) in formats that integrate cleanly with the broader scientific Python ecosystem (e.g., Xarray/Zarr) and with experiment tracking via MLflow. This makes it easier for researchers and applied scientists to share models and results, compare alternative specifications, and operationalize workflows on larger datasets than would be practical with bespoke scripts.

This seems reasonable to me, though it's only a potential future use.

The third is a statement about external adoption:
> Early external adoption is reflected in package index downloads: since its initial public release in June 2025, aimz has been downloaded over 10,000 times via PyPI and conda-forge as of January 2026.

I'm somewhat unsure how big of a number this is, or what would cross the threshold of "evidence of use beyond the authors' group." (For example, conda-forge downloads could be quite high if it was included in a Docker image, or something).

As this is a new JOSS criterion, I'm not sure how these judgement calls have been made so far.

---

All-in-all, I think the package's code looks nice, and is a convenient wrapper that gets rid of lots of boilerplate that shows up in numpyro code. Its engineering practices seem sound, and I would recommend acceptance barring the documentation and paper concerns.
