### 2026-02-23 bf067b3
[JOSS] Address reviewer comment on the paper:
1) Update bibliography entries
2) Enhance paper summary for clarity and accessibility

Fix #161
