+++ paper/paper.bib
@@ -48,12 +48,12 @@
   url       = {https://doi.org/10.5281/zenodo.3773450}
 }
 
-@article{Zaharia_Accelerating_the_Machine_2018,
+@article{zaharia2018accelerating,
   author  = {Zaharia, Matei A. and Chen, Andrew and Davidson, Aaron and Ghodsi, Ali and Hong, Sue Ann and Konwinski, Andy and Murching, Siddharth and Nykodym, Tomas and Ogilvie, Paul and Parkhe, Mani and Xie, Fen and Zumar, Corey},
   journal = {IEEE Data Eng. Bull.},
   pages   = {39--45},
   title   = {{Accelerating the Machine Learning Lifecycle with MLflow}},
-  url     = {https://api.semanticscholar.org/CorpusID:83459546},
+  url     = {https://dblp.org/rec/journals/debu/ZahariaCD0HKMNO18},
   volume  = {41},
   year    = {2018}
 }
