### 2026-02-26 a7e9196
- Enhance FAQ documentation on NumPyro model compatibility and clarify requirements for static shapes in aimz (Fixes #162).
- Update pre-commit hooks to latest versions.
