+++ docs/source/faq.rst
@@ -38,7 +38,7 @@ Fix it by renaming your arguments to ``X`` / ``y`` or supplying ``param_input`` 
 Do I need to know NumPyro_ to use aimz?
 ---------------------------------------
 Yes.
-aimz builds on NumPyro_’s primitives and effect handlers.
+The aimz package builds on NumPyro_’s primitives and effect handlers.
 You should be comfortable writing a model function, defining a guide (for SVI) or configuring MCMC, and reading model traces.
 The library focuses on orchestration, not abstracting away core probabilistic modeling concepts.
 
@@ -46,15 +46,96 @@ The library focuses on orchestration, not abstracting away core probabilistic mo
 Can I use aimz with any NumPyro_ model?
 ---------------------------------------
 No.
-Most conventional SVI / MCMC models with global latents and plate-based structure work, but there are limitations:
+Most conventional models with global latents and a plate-based structure work out of the box.
+The core requirement is that every sample site in the model must have a **static shape** that does not change with the data.
 
-* Use of the control‑flow primitive: :external:func:`~numpyro.contrib.control_flow.scan` (or deeply nested dynamic control flow) – only partial support.
-* Complex per‑example local latent structure (e.g., very large / nested plates) – only partial support.
-* Models that rely on shapes incompatible with sharded / batched execution may require refactoring.
+On multi-device systems, :meth:`~aimz.ImpactModel.predict` uses `data parallelism <https://jax.readthedocs.io/en/latest/distributed_data_loading.html#data-parallelism>`_: the input data is sharded across devices along axis 0 (the observation dimension), while posterior samples are **replicated** on every device.
+The forward sampler used internally also requires every traced site to have a fixed shape across all sample iterations.
 
-If a model traces successfully once but fails in batched prediction, try the
-``*_on_batch`` variants or simplify local latent structure.
-We plan to broaden coverage—if you hit an unsupported pattern (ideally with a minimal reproducible example), please open an issue or submit a PR.
+Several modeling patterns can violate this requirement.
+The two most common are:
+
+* **Local latent variables**
+
+  A :external:func:`~numpyro.primitives.sample` call inside a :external:class:`~numpyro.primitives.plate` without ``obs=`` produces a posterior whose shape grows with the plate size (e.g., ``(num_samples, n_obs)``).
+  Because ``n_obs`` can differ between training and prediction, the replicated-posterior contract breaks.
+
+* **The** :external:func:`~numpyro.contrib.control_flow.scan` **primitive**
+
+  :external:func:`~numpyro.contrib.control_flow.scan` is commonly used for sequential or autoregressive models (e.g., state-space models, time-series forecasting).
+  The number of sites it creates typically grows with the sequence length, making the trace shape dynamic for the same reason.
+
+When this incompatibility is detected, :meth:`~aimz.ImpactModel.predict` issues a warning and automatically falls back to :meth:`~aimz.ImpactModel.predict_on_batch`, which processes data in a single batch without sharding.
+However, if the posterior shapes are fundamentally incompatible with the new input, the forward pass will still fail with a shape mismatch.
+
+Note that a model with nested plates whose :external:func:`~numpyro.primitives.sample` sites are all observed (``obs=``) or whose latent shapes are fixed remains compatible.
+In contrast, a model with a single plate containing one unobserved :external:func:`~numpyro.primitives.sample` site triggers the fallback.
+The following examples illustrate the compatibility with :meth:`~aimz.ImpactModel.predict`:
+
+.. code-block:: python
+
+    import numpyro.distributions as dist
[…]
