### 2026-02-26 5dfa3ff
- Enhance FAQ documentation on NumPyro model compatibility and clarify requirements for static shapes in aimz (Fixes #162).
- Update pre-commit hooks to latest versions.
