### 2026-03-05 0640872
fix: prevent GPU OOM by deferring array transfer to batch time
- Pass `to_jax=False` in `_setup_inputs` so the full dataset stays as
  NumPy arrays instead of being eagerly moved to GPU via `jnp.asarray`.
- Place index arrays on CPU using `jax.default_device` to avoid
  unnecessary GPU memory usage for bookkeeping.
- Transfer data to GPU per-batch through the existing `device_put` path
  in `ArrayLoader.__iter__`.

  Fixes #168
