### 2026-03-06 f07a5e2
docs: enhance README and index.rst with detailed capabilities
