+++ docs/source/index.rst
@@ -9,13 +9,14 @@ aimz: Scalable probabilistic impact modeling
 `Issues <https://github.com/markean/aimz/issues>`__ |
 `Releases <https://github.com/markean/aimz/releases>`__
 
-aimz is a Python library for scalable probabilistic impact modeling, enabling assessment of intervention effects on outcomes with a streamlined interface for fitting, sampling, prediction, and effect estimation—minimal boilerplate, accelerated execution, and powered by `NumPyro <https://num.pyro.ai/en/stable/>`_, `JAX <https://jax.readthedocs.io/en/latest/>`_, `Xarray <https://xarray.dev/>`_, and `Zarr <https://zarr.readthedocs.io/en/stable/>`_.
-It features:
+aimz is a Python library for scalable probabilistic impact modeling—estimating how interventions affect outcomes while quantifying uncertainty. It provides an intuitive interface for fitting Bayesian models, drawing posterior samples, generating large-scale posterior predictive simulations, and estimating intervention effects with minimal boilerplate.
 
-* Intuitive API combining the ease of use from ML frameworks with the flexibility of probabilistic modeling.
-* Accelerated computation via parallelism and distributed data.
-* Support for interventional causal inference for counterfactuals and causal effects.
-* MLflow integration for experiment tracking and model management.
+* **Flexible model specification:** Built on `NumPyro <https://num.pyro.ai/en/stable/>`_ and `JAX <https://jax.readthedocs.io/en/latest/>`_, bring a NumPyro model as a "kernel"—aimz does not enforce a fixed architecture.
+* **Stochastic variational inference and MCMC:** Supports both SVI (including minibatch) and MCMC sampling through NumPyro's inference algorithms.
+* **Scalable predictive sampling:** JIT-compiled, sharded sampling streams results to chunked `Zarr <https://zarr.readthedocs.io/en/stable/>`_ stores, enabling large-scale posterior predictive simulations.
+* **Structured outputs:** Predictions, samples, and effect estimates are materialized as `Xarray <https://xarray.dev/>`_ objects backed by Zarr, integrating cleanly with the scientific Python ecosystem.
+* **Intervention handling:** Specify interventions declaratively and estimate effects from posterior predictive distributions.
+* **Experiment tracking:** `MLflow <https://mlflow.org/>`_ integration for logging runs, parameters, metrics, and model artifacts with full lineage.
 
 
 Navigation
