### 2026-03-06 913fd57
docs: enhance README and index.rst with detailed capabilities
