### 2026-03-06 293e176
refactor: replace JAX default_device with NumPy for index handling in ArrayLoader
