### 2026-03-09 3b3625c
refactor: update ArrayDataset default to_jax=False and fix docstrings
- Change ArrayDataset default from to_jax=True to to_jax=False
- Fix module, class, and method docstrings in array_dataset.py
- Remove unnecessary cast imports and calls in _input_setup.py
- Update test to reflect new default behavior
- Fix group_path docstring in _output.py

Fixes #170
