### 2026-03-09 104d759
fix: prevent zero batch size when num_samples exceeds threshold
Add test_input_setup.py to exercise the capping branch with a multi-device sharding.

Fixes #172
