### 2026-03-09 dc14ad7
fix: replace device_count with local_device_count
