### 2026-03-09 b43cbe3
refactor: use NumPy padding and skip pad when unnecessary
- pad_array now returns NumPy arrays when given NumPy input, avoiding premature GPU transfers before device_put
- Skip padding entirely when n_pad is 0 to avoid unnecessary copies
- Fix pad_array return type annotation (Array | npt.NDArray)
- Update module and class docstrings

Fixes #174
