### 2026-03-09 da4e5d9
fix: update changelog to include padding behavior for NumPy arrays in ArrayLoader
