### 2026-03-09 691e138
test: add JAX conversion test for ArrayDataset
