### 2026-03-09 #9738 @ankurankan
Hi, sorry again for the delay in the review. I have gone through the checklist and my comments are as follows:

> Open development: Was the software developed openly from early stages? For projects with recent public repositories, is there at least six months of public development history with evidence of releases, public issues/pull requests, and ideally external engagement?

The software has more than six months of open development history on github. However, I do not see any external engagement.

> Collaborative effort: Does the commit history show contributions from multiple developers and evidence of iterative refinement through community feedback?

I do not see any community involvement with the project.

>  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I do agree with @DanWaxman that a concrete real world example would be helpful.

Comments on the paper:

1. The "Statement of Need" section starts with describing problems with standard Bayesian workflows, but then switches
to causal inference toolkits. It is not clear how these are related in the context of aimz. More broadly, the paper
oscillates between presenting aimz as general-purpose Bayesian infrastructure and as a specialized tool for
interventional effect estimation. The Statement of Need reads broadly, the Research Impact Statement narrows to
intervention estimation, and the State of the Field mixes both. Clarifying the intended scope throughout would improve
readability.
2. The Statement of Need claims that "many causal inference toolkits focus on graph discovery." This is not accurate, as
many packages focus on the causal estimation task, such as EconML, causaleffects, DoubleML. I would suggest revising this claim to more
accurately reflect the landscape.
3. In the State of the Field section, the discussion of CausalPy feels out of place, as I am not able to connect it
directly to the Bayesian inference framing. If aimz is meant to be much more general, why the specific focus on these
causal packages?
4. I think it would be helpful to explain what an ImpactModel actually is. The paper mentions it as the central
abstraction but does not define it. A brief definition would be helpful.
5. The `estimate_effect` method carries strong causal connotations. I think it would be nice to clarify some of the
details such as how does this method handle the causal identification problem? What happens when the causal effect is
inestimable?
