### 2026-03-09 926f593
Add benchmark scripts and README for paper
