+++ paper/benchmark/bench_time_predictive.py
@@ -1,0 +1,179 @@
+"""Benchmark wall-clock time for NumPyro `Predictive`."""
+
+import argparse
+import logging
+import time
+import warnings
+from importlib.util import find_spec
+
+import jax
+import jax.numpy as jnp
+import numpyro.distributions as dist
+from flax import nnx
+from jax import (
+    Array,
+    block_until_ready,
+    default_backend,
+    default_device,
+    device_get,
+    devices,
+    local_device_count,
+    random,
+)
+from numpyro import plate, sample
+from numpyro.contrib.module import random_nnx_module
+from numpyro.infer import SVI, Predictive, Trace_ELBO, init_to_uniform
+from numpyro.infer.autoguide import AutoNormal
+from optax import adam
+
+if find_spec("flox") is None:
+    warnings.warn(
+        "The 'flox' package is required for efficient 'groupby' operations in Dask "
+        "arrays.",
+        stacklevel=2,
+    )
+
+# ── Constants ─────────────────────────────────────────────────────────────────────────
+NUM_FEATURES: int = 100
+D_HIDDEN: int = 128
+NUM_STEPS: int = 100
+NUM_SAMPLES: int = 1_000
+BATCH_SIZE: int = 10_000
+
+
+# ── Model ─────────────────────────────────────────────────────────────────────────────
+class MLP(nnx.Module):
+    """Four-layer MLP with ReLU activations and sigmoid output."""
+
+    dtype: jnp.dtype = jnp.bfloat16 if default_backend() == "gpu" else jnp.float32
+
+    def __init__(self, din: int, dmid: int, dout: int, *, rngs: nnx.Rngs) -> None:
+        """Initialize with input dim `din`, hidden dim `dmid`, and output dim `dout`."""
+        self.linear1 = nnx.Linear(din, dmid, dtype=self.dtype, rngs=rngs)
+        self.linear2 = nnx.Linear(dmid, dmid, dtype=self.dtype, rngs=rngs)
+        self.linear3 = nnx.Linear(dmid, dmid, dtype=self.dtype, rngs=rngs)
+        self.linear4 = nnx.Linear(dmid, dout, dtype=self.dtype, rngs=rngs)
+
+    def __call__(self, x: Array) -> Array:
+        """Forward pass returning sigmoid probabilities."""
+        x = self.linear1(x)
+        x = nnx.relu(x)
+        x = self.linear2(x)
+        x = nnx.relu(x)
+        x = self.linear3(x)
+        x = nnx.relu(x)
+        x = self.linear4(x)
+
+        return nnx.sigmoid(x).reshape(-1)
+
+
+# ── Helpers ───────────────────────────────────────────────────────────────────────────
+def generate_synthetic_data(rng_key: Array, n: int, p: int) -> tuple[Array, Array]:
+    """Generate synthetic data with a nonlinear data generating process.
+
+    Data is generated as JAX arrays on CPU because NumPyro's `Predictive` requires JAX
+    arrays. The arrays are later moved to the compute device via `device_put`.
+    """
+    with default_device(devices("cpu")[0]):
+        k1, k2, k3, k4 = random.split(rng_key, 4)
+        X = random.normal(k1, (n, p))
+        b1 = random.normal(k2, (p,))
+        b2 = random.normal(k3, (p,))
+        logits = jnp.sin(X @ b1) * jnp.tanh(X @ b2) + 0.5 * jnp.cos((X**2) @ b1)
+        y = random.bernoulli(k4, nnx.sigmoid(logits)).astype(jnp.int32)
+
+    return X, y
+
+
+# ── Main ──────────────────────────────────────────────────────────────────────────────
+logger = logging.getLogger(__name__)
+
+
[…]
