+++ paper/benchmark/bench_memory_predict.py
@@ -1,0 +1,234 @@
+"""Benchmark peak incremental RSS for aimz `predict`."""
+
+import argparse
+import contextlib
+import gc
+import logging
+import os
+import threading
+import warnings
+from collections.abc import Callable
+from importlib.util import find_spec
+from pathlib import Path
+
+import jax
+import jax.numpy as jnp
+import numpy as np
+import numpy.typing as npt
+import numpyro.distributions as dist
+from flax import nnx
+from jax import Array, default_backend, local_device_count, random
+from numpyro import plate, sample
+from numpyro.contrib.module import random_nnx_module
+from numpyro.infer import SVI, Trace_ELBO, init_to_uniform
+from numpyro.infer.autoguide import AutoNormal
+from optax import adam
+
+from aimz import ImpactModel
+from aimz.utils.data import ArrayDataset, ArrayLoader
+
+if find_spec("flox") is None:
+    warnings.warn(
+        "The 'flox' package is required for efficient 'groupby' operations in Dask "
+        "arrays.",
+        stacklevel=2,
+    )
+
+jax.config.update("jax_platform_name", "cpu")
+jax.config.update("jax_num_cpu_devices", 4)
+
+# ── Constants ─────────────────────────────────────────────────────────────────────────
+NUM_FEATURES: int = 100
+D_HIDDEN: int = 128
+NUM_STEPS: int = 100  # Fit quality is irrelevant.
+NUM_SAMPLES: int = 1_000
+BATCH_SIZE: int = 10_000
+
+
+# ── Model ─────────────────────────────────────────────────────────────────────────────
+class MLP(nnx.Module):
+    """Four-layer MLP with ReLU activations and sigmoid output."""
+
+    dtype: jnp.dtype = jnp.bfloat16 if default_backend() == "gpu" else jnp.float32
+
+    def __init__(self, din: int, dmid: int, dout: int, *, rngs: nnx.Rngs) -> None:
+        """Initialize with input dim `din`, hidden dim `dmid`, and output dim `dout`."""
+        self.linear1 = nnx.Linear(din, dmid, dtype=self.dtype, rngs=rngs)
+        self.linear2 = nnx.Linear(dmid, dmid, dtype=self.dtype, rngs=rngs)
+        self.linear3 = nnx.Linear(dmid, dmid, dtype=self.dtype, rngs=rngs)
+        self.linear4 = nnx.Linear(dmid, dout, dtype=self.dtype, rngs=rngs)
+
+    def __call__(self, x: Array) -> Array:
+        """Forward pass returning sigmoid probabilities."""
+        x = self.linear1(x)
+        x = nnx.relu(x)
+        x = self.linear2(x)
+        x = nnx.relu(x)
+        x = self.linear3(x)
+        x = nnx.relu(x)
+        x = self.linear4(x)
+
+        return nnx.sigmoid(x).reshape(-1)
+
+
+# ── RSS measurement ───────────────────────────────────────────────────────────────────
+def _current_rss_mib() -> float:
+    """Return current anonymous RSS in MiB."""
+    with Path.open(f"/proc/{os.getpid()}/status") as f:
+        for line in f:
+            if line.startswith("RssAnon:"):
+                return int(line.split()[1]) / 1024.0
+    msg = "RssAnon not found in /proc/self/status"
+    raise RuntimeError(msg)
+
+
+def _peak_rss_increment(fn: Callable) -> float:
+    """Run *fn()* and return the peak RSS increment."""
+    gc.collect()
+
+    baseline = _current_rss_mib()
+    peak = baseline
+    stop = threading.Event()
+
+    def _sampler() -> None:
+        nonlocal peak
+        while not stop.is_set():
[…]

[…]
