### 2026-03-09 c911275
Add benchmark scripts and README for paper
