+++ paper/paper.md
@@ -26,10 +26,6 @@ Researchers and analysts can explore “what-if” scenarios to understand poten
 `aimz` supports advanced inference methods with efficient, reproducible workflows, making it easy to run large-scale simulations and serialize structured outputs.
 These design choices reduce bespoke glue code, enable high-throughput analyses, and support rapid iteration and experimentation.
 
-It combines the usability of general machine learning APIs with the flexibility of probabilistic programming through a single high-level object (`ImpactModel`).
-Built atop JAX [@jax2018github] and NumPyro [@phan2019composable], it supports (minibatch) stochastic variational inference (SVI) and Markov chain Monte Carlo sampling, just-in-time (JIT)-compiled parallel predictive streaming to chunked Zarr [@alistair_miles_2020_3773450] stores exposed through Xarray [@hoyer2017xarray], and first-class intervention handling for effect estimation.
-Integrated MLflow [@zaharia2018accelerating] support enables experiment tracking and model lineage.
-
 # Statement of need
 
 Standard Bayesian workflows often encounter significant engineering friction when transitioning from model specification to large-scale impact evaluation.
@@ -43,7 +39,7 @@ This reduces the bespoke engineering effort required to connect flexible statist
 The landscape of Bayesian software is largely divided between low-level flexibility and high-level rigidity.
 Core probabilistic programming languages like NumPyro, PyMC [@pymc2023], and Stan [@carpenter2017stan] provide the flexible primitives necessary for custom research but require users to manually handle scaling, parallelization, and other performance considerations.
 Conversely, domain-specific frameworks like Meridian [@meridian_github], Robyn [@robyn], or PyMC-Marketing [@pymc-marketing] can offer robust end-to-end pipelines but are specialized for marketing mix modeling.
-These tools frequently enforce fixed model architectures (e.g., specific adstock or saturation transformations) and opaque internal logic that are difficult to adapt for broader scientific research, such as evaluating clinical care gaps---where individual-level discrepancies between recommended best practices and actual care require custom hierarchical specifications and structural modifications.
+These tools frequently enforce fixed model architectures (e.g., specific adstock or saturation transformations) and opaque internal logic that are difficult to adapt for broader scientific research, such as evaluating clinical care gaps, where individual-level discrepancies between recommended best practices and actual care require custom hierarchical specifications and structural modifications.
 
 `aimz` is designed to target a middle ground by providing Bayesian infrastructure in an estimator-like form.
 It does not prescribe a specific model form; instead, it provides a scalable execution layer for user-defined models.
@@ -71,9 +67,33 @@ The combination of a stable, method-based interface and standardized outputs mak
 # Research impact statement
 
 `aimz` enables a class of Bayesian modeling and probabilistic analyses that are otherwise costly to implement and difficult to reproduce at scale: including fitting flexible probabilistic models, generating large posterior predictive simulations, and estimating intervention effects under explicit structural modifications.
-Its research impact is therefore primarily infrastructural---lowering the engineering barrier to rigorous uncertainty-aware effect estimation---while remaining general enough to support diverse model specifications and inference strategies.
+Its research impact is therefore primarily infrastructural, lowering the engineering barrier to rigorous uncertainty-aware effect estimation while remaining general enough to support diverse model specifications and inference strategies.
 
-Within Eli Lilly and Company, `aimz` has supported internal analytics workflows that require scalable posterior and posterior predictive sampling, consistent output structures, and experiment traceability.
[…]
