### 2026-03-09 a0a2f81
doc: joss paper starts new paragraph after introducing benchmark settings
