### 2026-03-10 markean/aimz#177 @markean
Clarify scope of `estimate_effect` in documentation
The `estimate_effect` method name carries causal connotations that may lead users to expect built-in causal identification. The intervention guide should clarify that the method computes the posterior predictive contrast between two scenarios, and that whether the result admits a causal interpretation depends on the structural assumptions encoded in the user-defined model. A brief note on the scope would help set expectations for users approaching the method from a causal inference perspective.

Ref: https://github.com/openjournals/joss-reviews/issues/9738#issuecomment-4025947266.
