### 2026-03-10 9eae5da
doc: enhance intervention guide with causal interpretation notes and model compatibility reference
Fixes #177
