### 2026-03-10 e7023e6
doc: enhance intervention guide with causal interpretation notes and model compatibility reference
Fixes #177
