### 2026-03-10 ef9ad81
doc: address reviewer feedback on paper scope and framing
