+++ paper/paper.md
@@ -6,7 +6,7 @@ tags:
   - Bayesian inference
   - uncertainty quantification
   - probabilistic modeling
-  - uplift modeling
+  - high-performance computing
 authors:
   - name: Eunseop Kim
     orcid: 0009-0000-2138-788X
@@ -20,17 +20,17 @@ bibliography: paper.bib
 
 # Summary
 
-`aimz` is a Python library for probabilistic impact modeling, a method for estimating how actions or interventions affect outcomes while capturing uncertainty.
-It provides an intuitive interface for fitting Bayesian models, drawing posterior samples, generating large-scale posterior predictive simulations, and estimating interventional effects with minimal boilerplate.
-Researchers and analysts can explore “what-if” scenarios to understand potential effects under different decisions or interventions, generating predictions that reflect the full range of possible outcomes.
+`aimz` is a Python library that provides scalable infrastructure for Bayesian modeling and probabilistic simulation.
+It offers an intuitive, estimator-like interface for fitting Bayesian models, drawing posterior samples, generating large-scale posterior predictive simulations, and comparing outcomes across intervention scenarios with minimal boilerplate.
+Researchers and analysts can explore “what-if” scenarios to understand potential effects under different decisions, generating predictions that reflect the full range of possible outcomes.
 `aimz` supports advanced inference methods with efficient, reproducible workflows, making it easy to run large-scale simulations and serialize structured outputs.
 These design choices reduce bespoke glue code, enable high-throughput analyses, and support rapid iteration and experimentation.
 
 # Statement of need
 
-Standard Bayesian workflows often encounter significant engineering friction when transitioning from model specification to large-scale impact evaluation.
-While core probabilistic programming frameworks offer mature inference algorithms, they lack native infrastructure for the repetitive engineering tasks essential to production-grade impact analysis: streaming predictive draws without exceeding system memory, managing device-aware sharding for simulation, and coordinating complex interventional scenarios.
-General machine learning libraries lack the calibrated uncertainty quantification provided by Bayesian sampling, while many causal inference toolkits focus on graph discovery rather than high-throughput predictive workflows.
+Standard Bayesian workflows often encounter significant engineering friction when transitioning from model specification to large-scale predictive simulation and impact evaluation.
+While core probabilistic programming frameworks offer mature inference algorithms, they lack native infrastructure for the repetitive engineering tasks essential to production-grade analysis: streaming predictive draws without exceeding system memory, managing device-aware sharding for simulation, and coordinating complex interventional scenarios.
+General machine learning libraries provide scalable prediction pipelines but lack the calibrated uncertainty quantification provided by Bayesian sampling.
 `aimz` addresses these gaps by consolidating model tracing, sharded predictive sampling, and experiment lineage into a single estimator-like object.
 This reduces the bespoke engineering effort required to connect flexible statistical research with high-throughput data pipelines, supporting reliable and reproducible probabilistic analyses at scale.
 
@@ -38,19 +38,18 @@ This reduces the bespoke engineering effort required to connect flexible statist
 
 The landscape of Bayesian software is largely divided between low-level flexibility and high-level rigidity.
 Core probabilistic programming languages like NumPyro, PyMC [@pymc2023], and Stan [@carpenter2017stan] provide the flexible primitives necessary for custom research but require users to manually handle scaling, parallelization, and other performance considerations.
-Conversely, domain-specific frameworks like Meridian [@meridian_github], Robyn [@robyn], or PyMC-Marketing [@pymc-marketing] can offer robust end-to-end pipelines but are specialized for marketing mix modeling.
[…]
