+++ paper/paper.bib
@@ -109,13 +109,6 @@
   year    = {2025}
 }
 
-@software{causalpy_github,
-  title   = {CausalPy: Causal inference for quasi-experiments in Python},
-  author  = {{PyMC Labs}},
-  url     = {https://github.com/pymc-labs/CausalPy},
-  version = {0.7.0},
-  year    = {2026}
-}
 
 @software{robyn,
   title   = {Robyn: Semi-Automated Marketing Mix Modeling (MMM) from Meta Marketing Science},
@@ -132,4 +125,15 @@
   url     = {https://github.com/pymc-labs/pymc-marketing},
   version = {0.16.0},
   year    = {2025}
+}
+
+@article{prophet,
+  title   = {Forecasting at scale},
+  author  = {Taylor, Sean J. and Letham, Benjamin},
+  journal = {The American Statistician},
+  volume  = {72},
+  number  = {1},
+  pages   = {37--45},
+  year    = {2018},
+  doi     = {10.1080/00031305.2017.1380080}
 }
