### 2026-03-10 4201c6d
doc: revise wordings to use the term Bayesian inference
