### 2026-03-10 476b838
doc: introduce reference for NumPyro when it is first used in the paper
