### 2026-03-10 b3ba39f
doc: simplify description of ImpactModel functionality
