+++ paper/paper.md
@@ -49,7 +49,7 @@ This makes `aimz` suited for production-grade Bayesian workflows where reproduci
 
 `aimz` is designed to support scalable Bayesian analyses in applied settings, allowing users to iterate quickly across model specifications, inference settings, and intervention scenarios, as well as to handle large datasets and produce reproducible artifacts.
 It combines the usability of general machine learning APIs with the flexibility of probabilistic programming through a single high-level object (`ImpactModel`).
-An `ImpactModel` wraps a user-defined NumPyro model function — its *kernel* — together with an inference strategy and a JAX random key, and exposes methods for fitting, prediction, and scenario comparison that manage batching, sharding, and artifact serialization internally.
+An `ImpactModel` wraps a user-defined NumPyro model function together with an inference strategy, and exposes methods for fitting, prediction, and scenario comparison that manage batching, sharding, and artifact serialization internally.
 Built atop JAX [@jax2018github] and NumPyro, `aimz` leverages modeling flexibility, accelerator-native execution, and composable program transformations such as vectorization and sharded execution.
 It supports (minibatch) stochastic variational inference (SVI) and Markov chain Monte Carlo sampling, just-in-time–compiled parallel predictive streaming to chunked Zarr [@alistair_miles_2020_3773450] stores exposed through Xarray [@hoyer2017xarray], and first-class intervention handling for effect estimation.
 Integrated MLflow [@zaharia2018accelerating] support enables experiment tracking and model lineage.
