### 2026-03-10 b6d87f3
doc: remove unnecessary quotation marks from "glue" code description in paper
