### 2026-03-12 651e1c0
Add on_batch parameter to estimate_effect
Allow `estimate_effect()` to dispatch predictions through `predict_on_batch` via the new `on_batch` keyword-only argument. When the underlying method returns a raw dict instead of a `DataTree`, the result is automatically wrapped in a `DataTree` with the appropriate group node

Widen `_dict_to_datatree()` to accept `Mapping[str, Array | NDArray]`
so both JAX and NumPy arrays are handled.

Also reorder ImpactModel properties alphabetically, expose `inference`
as a read-only property, and fix NumPyro docstring cross-references.

Fixes #180.
