+++ aimz/model/impact_model.py
@@ -98,10 +98,10 @@ class ImpactModel(BaseModel):
         """Initialize an :class:`~aimz.ImpactModel` instance.
 
         Args:
-            kernel: A probabilistic model with NumPyro primitives.
+            kernel: A probabilistic model with `NumPyro`_ primitives.
             rng_key: A pseudo-random number generator key.
-            inference: An inference method supported by NumPyro, such as an instance of
-                :external:class:`~numpyro.infer.svi.SVI` or
+            inference: An inference method supported by `NumPyro`_, such as an instance
+                of :external:class:`~numpyro.infer.svi.SVI` or
                 :external:class:`~numpyro.infer.mcmc.MCMC`.
             param_input: Name of the parameter in the ``kernel`` for the main input
                 data.
@@ -125,7 +125,7 @@ class ImpactModel(BaseModel):
                 "Expected `SVI` or `MCMC` from `numpyro.infer`."
             )
             raise TypeError(msg)
-        self.inference = inference
+        self._inference = inference
         self._vi_result: SVIRunResult | None = None
         self._vi_state = None
         self._posterior: dict[str, Array] | None = None
@@ -230,19 +230,19 @@ class ImpactModel(BaseModel):
         self._init_runtime_attrs()
 
     @property
+    def inference(self) -> SVI | MCMC:
+        """The underlying `NumPyro`_ inference object."""
+        return self._inference
+
+    @property
     def kernel(self) -> Callable:
-        """A probabilistic model with NumPyro primitives."""
+        """A probabilistic model with `NumPyro`_ primitives."""
         return self._kernel
 
     @property
     def kernel_spec(self) -> KernelSpec | None:
         """The cached :class:`~aimz.model.KernelSpec` or ``None`` if not yet built."""
         return self._kernel_spec
-
-    @property
-    def rng_key(self) -> Array:
-        """Pseudo-random number generator key."""
-        return self._rng_key
 
     @property
     def param_input(self) -> str:
@@ -253,6 +253,21 @@ class ImpactModel(BaseModel):
     def param_output(self) -> str:
         """Parameter name in :attr:`~aimz.ImpactModel.kernel` for the output data."""
         return self._param_output
+
+    @property
+    def posterior(self) -> dict[str, Array] | None:
+        """Posterior samples by variable name, or ``None`` if not set."""
+        return self._posterior
+
+    @property
+    def rng_key(self) -> Array:
+        """Pseudo-random number generator key."""
+        return self._rng_key
+
+    @property
+    def temp_dir(self) -> str | None:
+        """Temporary directory path, or ``None`` if not set."""
+        return self._temp_dir.name if self._temp_dir else None
 
     @property
     def vi_result(self) -> SVIRunResult | None:
@@ -282,16 +297,6 @@ class ImpactModel(BaseModel):
         self._is_fitted = True
 
         self._vi_result = vi_result
-
-    @property
-    def posterior(self) -> dict[str, Array] | None:
-        """Posterior samples by variable name, or ``None`` if not set."""
-        return self._posterior
-
-    @property
-    def temp_dir(self) -> str | None:
-        """Temporary directory path, or ``None`` if not set."""
-        return self._temp_dir.name if self._temp_dir else None
 
     def _build_kernel_spec(
         self,
@@ -424,7 +429,7 @@ class ImpactModel(BaseModel):
             rng_key: Pseudo-random number generator key.
             return_sites: Names of variables (sites) to return.
             output_dir: Directory where outputs will be saved.
-            kernel: Probabilistic model with NumPyro primitives.
+            kernel: Probabilistic model with `NumPyro`_ primitives.
             sampler: Function performing predictive sampling; must accept the same
                 signature as this function.
             samples: Arrays to condition predictions on.
[…]
