+++ pyproject.toml
@@ -38,13 +38,15 @@ dev = ["pre-commit>=4.5", "pytest>=9", "pytest-cov>=7", "ruff>=0.15"]
 gpu = ["jax[cuda13]>=0.9"]
 mlflow = ["mlflow>=3.5"]
 docs = [
-    "arviz-plots[matplotlib]>=0.6",
+    "arviz_base>=1; python_version >= '3.12'",
+    "arviz_plots>=1; python_version >= '3.12'",
+    "arviz_stats>=1; python_version >= '3.12'",
     "flax>=0.12",
     "jupyter-sphinx>=0.5",
     "mlflow>=3.5",
     "myst-parser>=5",
     "pydata-sphinx-theme>=0.16",
-    "sphinx>=8.2",
+    "sphinx>=9",
     "sphinx-copybutton>=0.5",
     "sphinx-design>=0.7",
 ]
