+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- {meth}`~aimz.ImpactModel.estimate_effect` now accepts an `on_batch` keyword argument.
+When ``True``, predictions are dispatched through {meth}`~aimz.ImpactModel.predict_on_batch` and any raw `dict` results are automatically converted to {class}`xarray.DataTree` ([#180](https://github.com/markean/aimz/issues/180)).
+
 ### Changed
 
 - {class}`~aimz.utils.data.ArrayDataset` now employs NumPy-based indexing in {class}`~aimz.utils.data.ArrayLoader` instead of triggering JAX tracing on each batch ([#168](https://github.com/markean/aimz/issues/168)).
