### 2026-03-12 c1e25af
test: add validation for dict results in `estimate_effect()` with `on_batch`
