+++ docs/source/examples/index.rst
@@ -1,0 +1,12 @@
+Examples
+========
+
+These examples demonstrate aimz's core capabilities across a range of real-world datasets.
+Each example highlights a different aspect of the :class:`~aimz.ImpactModel` API.
+
+.. toctree::
+   :maxdepth: 1
+
+   hillstrom
+   lalonde
+   criteo
