+++ docs/source/examples/hillstrom.rst
@@ -1,0 +1,724 @@
+Uplift Modeling with Custom Likelihood
+======================================
+
+This example uses the `Hillstrom email marketing dataset
+<https://blog.minethatdata.com/2008/03/minethatdata-e-mail-analytics-and-data.html>`_
+to estimate the causal effect of two email campaigns on customer conversion and
+spending.
+We build two `NumPyro`_ models, one for conversion and one for spend (a logistic
+regression and a hurdle model with a custom likelihood via
+:external:func:`~numpyro.primitives.factor`, respectively), fit them through the
+:class:`~aimz.ImpactModel` interface, and use :meth:`~aimz.ImpactModel.estimate_effect`
+to compute treatment effects.
+
+.. jupyter-execute::
+    :hide-output:
+
+    import arviz_base as az
+    import arviz_plots as azp
+    import arviz_stats as azs
+    import jax
+    import jax.numpy as jnp
+    import matplotlib.pyplot as plt
+    import matplotlib.ticker as mtick
+    import numpy as np
+    import numpyro.distributions as dist
+    import pandas as pd
+    from jax import Array, random
+    from numpyro import deterministic, factor, plate, sample
+    from numpyro.infer import MCMC, NUTS
+
+    from aimz import ImpactModel
+
+    # Force JAX to use CPU even if GPU is available
+    jax.config.update("jax_platform_name", "cpu")
+    # Set the number of CPU devices JAX sees (for CPU-based parallelism)
+    jax.config.update("jax_num_cpu_devices", 2)
+
+    # Configure the inline backend for high-resolution figures
+    %config InlineBackend.figure_format = "retina"
+
+    # Set the style for ArviZ plots
+    azp.style.use("arviz-variat")
+
+    # Set a random seed for reproducibility
+    rng_key = random.key(532)
+
+
+The Hillstrom Dataset
+---------------------
+
+The dataset comes from a randomized experiment at an e-commerce company, where 64,000
+customers who had made a purchase within the past twelve months were randomly assigned
+to one of three groups:
+
+- **Men's E-Mail**: received an email promoting men's merchandise.
+- **Women's E-Mail**: received an email promoting women's merchandise.
+- **No E-Mail**: control group, received no email.
+
+Two weeks after the email was sent, three outcomes were recorded: whether the customer
+visited the website, whether they made a purchase, and how much they spent.
+
+.. jupyter-execute::
+
+    df = pd.read_csv(
+        "http://www.minethatdata.com/"
+        "Kevin_Hillstrom_MineThatData_E-MailAnalytics"
+        "_DataMiningChallenge_2008.03.20.csv",
+    )
+
+\
+
+The dataset contains the following 11 columns:
+
+- **Covariates** (pre-treatment customer attributes):
+
+  - ``recency``: months since last purchase (integer, 1--12).
+  - ``history``: total dollar value spent in the past year (continuous).
+  - ``mens``: 1 if the customer purchased men's merchandise in the past year, 0
+    otherwise.
+  - ``womens``: 1 if the customer purchased women's merchandise in the past year, 0
+    otherwise.
+  - ``newbie``: 1 if the customer is new (first purchase within twelve months), 0
+    otherwise.
+  - ``zip_code``: residential area classification (Urban / Suburban / Rural).
+  - ``channel``: purchase channel in the past year (Phone / Web / Multichannel).
+
+- **Treatment**: ``segment``, one of "Mens E-Mail", "Womens E-Mail", or "No E-Mail".
+
[…]
