+++ docs/source/examples/lalonde.rst
@@ -1,0 +1,492 @@
+Effect Estimation with Intervention
+====================================
+
+This example uses the `Lalonde job training dataset
+<https://users.nber.org/~rdehejia/nswdata2.html>`_, a classic benchmark for causal
+inference with observational confounding, to estimate the effect of a job training
+program on earnings.
+The model encodes treatment as a `NumPyro`_ :func:`~numpyro.primitives.sample` site and
+uses the ``intervention`` keyword to apply `NumPyro`_'s
+:external:class:`~numpyro.handlers.do` handler, fixing treatment to specific values and
+generating counterfactual predictions without rewriting the model.
+
+The model includes a treatment x covariate interaction, decomposing the overall
+**average treatment effect** (ATE) into subgroup-specific **conditional average
+treatment effects** (CATEs) to ask whether the program helped those without a
+high-school degree more than those with one.
+
+.. jupyter-execute::
+
+    import arviz_base as az
+    import arviz_plots as azp
+    import arviz_stats as azs
+    import jax
+    import jax.numpy as jnp
+    import matplotlib.pyplot as plt
+    import numpy as np
+    import numpyro.distributions as dist
+    import pandas as pd
+    from jax import Array, random
+    from numpyro import deterministic, plate, sample
+    from numpyro.infer import MCMC, NUTS
+
+    from aimz import ImpactModel
+
+    # Force JAX to use CPU even if GPU is available
+    jax.config.update("jax_platform_name", "cpu")
+    # Set the number of CPU devices JAX sees (for CPU-based parallelism)
+    jax.config.update("jax_num_cpu_devices", 2)
+
+    # Configure the inline backend for high-resolution figures
+    %config InlineBackend.figure_format = "retina"
+
+    # Set the style for ArviZ plots
+    azp.style.use("arviz-variat")
+
+    # Set a random seed for reproducibility
+    rng_key = random.key(532)
+
+
+The Lalonde Dataset
+-------------------
+
+The dataset comes from an observational study of a job training program.
+A total of 614 individuals were observed, of whom 185 received training and 429 did not.
+The outcome is earnings in 1978 (``re78``), measured in thousands of dollars.
+
+.. code-block:: python
+
+    url = (
+        "https://raw.githubusercontent.com/robjellis/lalonde/"
+        "master/lalonde_data.csv"
+    )
+    df = pd.read_csv(url)
+
+    # Scale earnings to $k
+    df["re75"] = df["re75"] / 1_000
+    df["re78"] = df["re78"] / 1_000
+
+.. jupyter-execute::
+    :hide-code:
+    :hide-output:
+
+    from io import StringIO
+    import warnings
+
+    import requests
+    from urllib3.exceptions import InsecureRequestWarning
+
+    warnings.filterwarnings("ignore", category=InsecureRequestWarning)
+
+    url = (
+        "https://raw.githubusercontent.com/robjellis/lalonde/"
+        "master/lalonde_data.csv"
+    )
+    r = requests.get(url, verify=False)
+    df = pd.read_csv(StringIO(r.text))
+
+    # Scale earnings to $k
+    df["re75"] = df["re75"] / 1_000
+    df["re78"] = df["re78"] / 1_000
+
+\
+
+The dataset contains the following columns:
+
+- **Covariates** (pre-treatment):
+
+  - ``educ``: years of education.
+  - ``age``: age in years.
+  - ``re75``: earnings in 1975, in thousands of dollars.
+  - ``black``, ``hispan``: race/ethnicity indicators.
[…]
