### 2026-03-13 7e62951
Add documentation for examples and refactor for quality improvement:
- Created a new index for examples showcasing aimz's core capabilities.
- Added a detailed examples with real world datasets.
- Removed the getting started section and its related installation and tutorial files.
- Introduced a new installation guide in the user guide with detailed installation.
- Cleaned up various references to external libraries in the user guide.

Fixes #163
