+++ docs/source/examples/criteo.rst
@@ -1,0 +1,410 @@
+Scalable Predictive Sampling
+============================
+
+This example uses the `Criteo Uplift Prediction dataset
+<https://ailab.criteo.com/criteo-uplift-prediction-dataset/>`_,
+a large-scale advertising benchmark with approximately 14 million observations, to
+demonstrate aimz's scalable inference pipeline: minibatch stochastic variational
+inference (SVI) with :meth:`~aimz.ImpactModel.fit`, streaming posterior predictive
+sampling to Zarr_ with :meth:`~aimz.ImpactModel.predict`, and deterministic resource
+cleanup with :meth:`~aimz.ImpactModel.cleanup`.
+
+The model is a Bayesian neural network (BNN) built with
+`Flax <https://flax.readthedocs.io/>`_ and lifted into NumPyro_ through
+:external:func:`~numpyro.contrib.module.random_nnx_module`.
+Treatment is passed as an input feature, and the :class:`~aimz.ImpactModel` API handles
+batched training and prediction without requiring changes to the model code.
+
+.. code-block:: python
+
+    import jax
+    import jax.numpy as jnp
+    import matplotlib.pyplot as plt
+    import numpy as np
+    import numpyro.distributions as dist
+    import pandas as pd
+    from flax import nnx
+    from jax import Array, default_backend, random
+    from jax.typing import ArrayLike
+    from numpyro import deterministic, plate, sample
+    from numpyro.contrib.module import random_nnx_module
+    from numpyro.infer import SVI, Trace_ELBO, init_to_uniform
+    from numpyro.infer.autoguide import AutoNormal
+    from optax import adam
+
+    from aimz import ImpactModel
+
+    # Configure the inline backend for high-resolution figures
+    %config InlineBackend.figure_format = "retina"
+
+    # Set a random seed for reproducibility
+    rng_key = random.key(532)
+
+
+The Criteo Uplift Dataset
+-------------------------
+
+The Criteo Uplift Prediction dataset (v2.1) was constructed from several randomized
+incrementality tests in online advertising.
+In each test, a random subset of users was withheld from ad targeting (control) while
+the remainder was eligible for ads (treated).
+The treatment group is deliberately large (~85%) because withholding ads from control
+users has a direct revenue cost.
+The dataset contains approximately 14 million rows, each corresponding to a user with 12
+anonymized features (``f0`` through ``f11``), a binary treatment indicator, and two
+binary outcomes.
+The dataset is also available from
+`Hugging Face <https://huggingface.co/datasets/criteo/criteo-uplift>`_
+(~300 MB compressed).
+
+.. code-block:: python
+
+    df = pd.read_parquet("criteo.parquet")
+
+The dataset contains the following columns:
+
+- **Features**: ``f0`` through ``f11``, 12 dense float-valued covariates.
+  Feature names are anonymized and values are randomly projected to preserve
+  predictive power while preventing recovery of the original user context.
+
+- **Treatment**: ``treatment``, binary indicator (1 = user was eligible for
+  ad targeting, 0 = withheld from targeting).  Note that ``treatment = 1``
+  does not guarantee the user saw an ad; it means they were eligible for
+  targeting in the real-time bidding auction.
+
+- **Exposure**: ``exposure``, binary indicator of whether the user was
+  actually shown an ad.  By construction, ``treatment = 0`` implies
[…]
