+++ docs/source/getting_started/index.rst
@@ -1,8 +1,0 @@
-Getting Started
-===============
-
-.. toctree::
-   :maxdepth: 1
-
-   installation
-   tutorial
