### 2026-03-13 #9738 @markean
Hi @DanWaxman and @ankurankan,

Apologies for the delay in responding, and thank you for taking the time to review the package. I’ve addressed your comments in the updated paper and documentation.

Major updates include:

1. **Revised paper scope**: The paper now focuses on aimz as a Bayesian infrastructure layer rather than a causal inference toolkit. Since the initial submission, the scope and development direction have shifted internally. Some naming conventions (e.g., estimate_effect, ImpactModel) reflect the original causal focus and are retained as it still remains in scope for future development.
2. **New examples section**: The documentation now includes a dedicated [Examples](https://aimz.readthedocs.io/latest/examples/index.html) section with three end-to-end analyses on real datasets, each highlighting a different aspect of the ImpactModel API.
3. **Benchmark section**: The paper now includes a benchmark comparing predictive sampling throughput, with reproducible code provided. This addresses the updated JOSS review checklist requirement and contributes to the Research Impact Statement.

Regarding the specific comments from @ankurankan:

> Hi, sorry again for the delay in the review. I have gone through the checklist and my comments are as follows:
>
> > Open development: Was the software developed openly from early stages? For projects with recent public repositories, is there at least six months of public development history with evidence of releases, public issues/pull requests, and ideally external engagement?
>
> The software has more than six months of open development history on github. However, I do not see any external engagement.
>
> > Collaborative effort: Does the commit history show contributions from multiple developers and evidence of iterative refinement through community feedback?
>
> I do not see any community involvement with the project.
>
> > Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>
> I do agree with [@DanWaxman](https://github.com/DanWaxman) that a concrete real world example would be helpful.
>
> Comments on the paper:
>
> 1. The "Statement of Need" section starts with describing problems with standard Bayesian workflows, but then switches
>    to causal inference toolkits. It is not clear how these are related in the context of aimz. More broadly, the paper
>    oscillates between presenting aimz as general-purpose Bayesian infrastructure and as a specialized tool for
>    interventional effect estimation. The Statement of Need reads broadly, the Research Impact Statement narrows to
>    intervention estimation, and the State of the Field mixes both. Clarifying the intended scope throughout would improve
>    readability.
> 2. The Statement of Need claims that "many causal inference toolkits focus on graph discovery." This is not accurate, as
>    many packages focus on the causal estimation task, such as EconML, causaleffects, DoubleML. I would suggest revising this claim to more
>    accurately reflect the landscape.
> 3. In the State of the Field section, the discussion of CausalPy feels out of place, as I am not able to connect it
>    directly to the Bayesian inference framing. If aimz is meant to be much more general, why the specific focus on these
>    causal packages?
> 4. I think it would be helpful to explain what an ImpactModel actually is. The paper mentions it as the central
>    abstraction but does not define it. A brief definition would be helpful.
> 5. The `estimate_effect` method carries strong causal connotations. I think it would be nice to clarify some of the
>    details such as how does this method handle the causal identification problem? What happens when the causal effect is
>    inestimable?

Comments 1–4 have been addressed in the revised paper. Regarding comment 5, causal identification is outside the scope of this package. To help clarify this, I’ve added a brief note in the [user guide](https://aimz.readthedocs.io/latest/user_guide/intervention.html#effect-estimation).

I'm happy to discuss any of these changes further. Please let me know if there are remaining items on your checklists.
