### 2026-03-19 #9738 @DanWaxman
Hi! I didn't see anywhere that "evidence of use beyond the authors" was discussed or argued. That was my main remaining point in the checklist, outside of GitHub issues that the author has addressed.
