### 2026-03-19 #9738 @markean
Hi @DanWaxman!, thank you for raising this.

aimz does not yet have evidence of external use beyond the authors. To address this checklist item, the paper relied on the "credible near-term significance" path described in the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#research-impact-statement-required):

> "Evidence of realized impact (publications, external use, integrations) **or** credible near-term significance (benchmarks, reproducible materials, community-readiness signals)."

To this end, a comparative benchmark of predictive sampling throughput and three end-to-end examples on real datasets were added in this review, along with updates to the API documentation, tests, and contribution guidelines.

@DanWaxman @rich2355, please let me know if this addresses the checklist item or if additional evidence is needed.
