### 2026-03-22 #9738 @rich2355
> [@markean](https://github.com/markean) [@rich2355](https://github.com/rich2355) Looks good to me. BTW, I am not sure why the editorial bot is not generating the updated paper.

If this happened, the reason probably is the author may not modify the paper.md file, or not merge this file in the branch used for JOSS review. @markean lmk if you are able to fix this, and if not, I can consult with the team. Thanks!
