### 2026-03-22 #9738 @rich2355
> Hi! I didn't see anywhere that "evidence of use beyond the authors" was discussed or argued. That was my main remaining point in the checklist, outside of GitHub issues that the author has addressed.

@markean This should belong to the "Research Impact Statement" section that is required in the paper. Just check whether you are aware of this. If you have done that, please ignore my message. Lmk if there are any other questions. Thanks!
