### 2026-03-22 #9738 @markean
Hi @rich2355, @ankurankan, apologies for the confusion. The updates were on the main branch but hadn’t been merged into the joss branch. This has now been fixed, and the build is picking it up correctly  (including the updated benchmark results in the Research Impact Statement section).
