+++ .github/workflows/coverage.yaml
@@ -38,6 +38,6 @@ jobs:
         run: uv run pytest tests --cov=aimz --cov-report=xml --cov-report=term
 
       - name: Upload coverage reports to Codecov
-        uses: codecov/codecov-action@v5
+        uses: codecov/codecov-action@v6
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
