### 2026-03-31 75da98d
chore: update Python version support and dependencies in configuration files
Fixes #121
