### 2026-03-31 7673707
chore: update Python version support and dependencies in configuration files
Fixes #121
