+++ pyproject.toml
@@ -15,11 +15,11 @@ authors = [
 maintainers = [{ name = "Eunseop Kim", email = "markean@pm.me" }]
 requires-python = ">=3.11"
 dependencies = [
-    "dask>=2025.7",
+    "dask>=2025.10",
     "jax>=0.9",
-    "numpyro>=0.19.0",
+    "numpyro>=0.20.0",
     "scikit-learn>=1.6.1",
-    "xarray>=2025.7",
+    "xarray>=2025.10",
     "zarr>=3,<4",
 ]
 classifiers = [
@@ -31,6 +31,7 @@ classifiers = [
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
 ]
 
 [project.optional-dependencies]
