### 2026-03-31 58eaf6d
chore: add Python 3.14 to the Ruff workflow matrix
