### 2026-04-07 #9738 @DanWaxman
Hi! I've checked the remaining items, with the following two caveats:

1. Regarding collaborative effort, I have checked not because I found the item fulfilled, but because the AE suggested this was okay for this submission.
2. I'm still not 100% sure about research impact. The guidelines are somewhat vague to me. The benchmarks in the revised document make it clear that there is a use case, and there are plausible users, though I don't think this suggests current or necessarily future widespread adoption.
