### 2026-04-07 #9738 @rich2355
> Hi! I've checked the remaining items, with the following two caveats:
>
> 1. Regarding collaborative effort, I have checked not because I found the item fulfilled, but because the AE suggested this was okay for this submission.
> 2. I'm still not 100% sure about research impact. The guidelines are somewhat vague to me. The benchmarks in the revised document make it clear that there is a use case, and there are plausible users, though I don't think this suggests current or necessarily future widespread adoption.

I think this package is more like a specialized project, therefore, it makes sense to me that in the short term there will be no widespread adoption. That said, if you feel this project has a high risk of "having no impact in the long term", I may need to ask the author @markean to provide more information, and I will also consult with the team to see whether the author's response is sufficient.
