### 2026-04-10 #9738 @rich2355
> Hi! I guess it's unclear to me what "having no impact in the long term" actually means. I think people in this niche are likely to find and use the software, if that counts. This niche is not huge, but this software is certainly useful for people with industry-scale probabilistic model pipelines.

Yes, from my understanding, this is a confirmation that this project has research impact.
