### 2026-04-17 #9738 @rich2355
Hi @markean, for the six papers without DOIs, I am simply checking whether they are indeed missing DOIs. If yes, just let me know and I will proceed. Otherwise, please fix them. Thanks!
