### 2026-04-18 68b5b38
chore: clean up bibliography entries and update formatting
