+++ paper/paper.bib
@@ -1,139 +1,104 @@
-@software{jax2018github,
-  author  = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
-  title   = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
-  url     = {http://github.com/jax-ml/jax},
-  version = {0.3.13},
-  year    = {2018}
-}
-@article{phan2019composable,
-  title   = {Composable Effects for Flexible and Accelerated Probabilistic Programming in NumPyro},
-  author  = {Phan, Du and Pradhan, Neeraj and Jankowiak, Martin},
-  journal = {arXiv preprint arXiv:1912.11554},
-  year    = {2019}
-}
-
 @software{alistair_miles_2020_3773450,
-  author    = {Alistair Miles and
-               John Kirkham and
-               Martin Durant and
-               James Bourbeau and
-               Tarik Onalan and
-               Joe Hamman and
-               Zain Patel and
-               shikharsg and
-               Matthew Rocklin and
-               raphael dussin and
-               Vincent Schut and
-               Elliott Sales de Andrade and
-               Ryan Abernathey and
-               Charles Noyes and
-               sbalmer and
-               pyup.io bot and
-               Tommy Tran and
-               Stephan Saalfeld and
-               Justin Swaney and
-               Josh Moore and
-               Joe Jevnik and
-               Jerome Kelleher and
-               Jan Funke and
-               George Sakkis and
-               Chris Barnes and
-               Anderson Banihirwe},
-  title     = {zarr-developers/zarr-python: v2.4.0},
-  month     = jan,
-  year      = 2020,
-  publisher = {Zenodo},
-  version   = {v2.4.0},
-  doi       = {10.5281/zenodo.3773450},
-  url       = {https://doi.org/10.5281/zenodo.3773450}
+	title        = {zarr-developers/zarr-python: v2.4.0},
+	author       = {Alistair Miles and John Kirkham and Martin Durant and James Bourbeau and Tarik Onalan and Joe Hamman and Zain Patel and shikharsg and Matthew Rocklin and raphael dussin and Vincent Schut and Elliott Sales de Andrade and Ryan Abernathey and Charles Noyes and sbalmer and pyup.io bot and Tommy Tran and Stephan Saalfeld and Justin Swaney and Josh Moore and Joe Jevnik and Jerome Kelleher and Jan Funke and George Sakkis and Chris Barnes and Anderson Banihirwe},
+	year         = {2020},
+	publisher    = {Zenodo},
+	doi          = {10.5281/zenodo.3773450},
+	url          = {https://doi.org/10.5281/zenodo.3773450},
+	version      = {v2.4.0}
 }
-
-@article{zaharia2018accelerating,
-  author  = {Zaharia, Matei A. and Chen, Andrew and Davidson, Aaron and Ghodsi, Ali and Hong, Sue Ann and Konwinski, Andy and Murching, Siddharth and Nykodym, Tomas and Ogilvie, Paul and Parkhe, Mani and Xie, Fen and Zumar, Corey},
-  journal = {IEEE Data Eng. Bull.},
-  pages   = {39--45},
-  title   = {{Accelerating the Machine Learning Lifecycle with MLflow}},
-  url     = {https://dblp.org/rec/journals/debu/ZahariaCD0HKMNO18},
-  volume  = {41},
-  year    = {2018}
[…]
