### 2026-04-18 5caff25
Update dependencies and classifiers in pyproject.toml and uv.lock
- Restrict jax version to <0.10 in both pyproject.toml and uv.lock
- Update development status classifier from Alpha to Beta in pyproject.toml
- Upgrade pandas version from 2.3.2 to 2.3.3 in uv.lock, including updated source and wheel URLs
