### 2026-04-18 #9738 @rich2355
Thanks @markean, please also

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> Archive the release on Zenodo/figshare/etc and post the DOI here.

Lmk these information so that I can proceed. Thanks!
