### 2026-04-18 fa0ebf4
Bump version from 0.10.0.dev0 to 0.10.0 in pyproject.toml and uv.lock
