### 2026-04-18 #9738 @markean
@rich2355, I've released a new version (v0.10.0) with DOI: https://doi.org/10.5281/zenodo.19638071.

Thank you!
