### 2026-04-18 #9738 @rich2355
Hi @markean, please ensure that **everything** (the title, author list, version tag, and the license) in the joss paper is the same as what is in Zonodo. Right now I find that at least your authorship is inconsistent. Please check and correct and I will move forward to recommending acceptance. Thanks!
