### 2026-04-18 #9738 @markean
Hi @rich2355, I have made the necessary updates in Zenodo https://zenodo.org/records/19638071. The title, author list, version tag, and license now match those in the paper. Thank you!
